% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_pos.R
\name{adjust_positions}
\alias{adjust_positions}
\title{Adjust measurement positions in NMR dataset}
\usage{
adjust_positions(data, B_ref, B_new, bw_ref, bw_new, dp_ref, dp_new)
}
\arguments{
\item{data}{matrix of NMR signals to be adjusted. Each row contains one spectrum.}

\item{B_ref}{numeric, magnetic field strength of reference}

\item{B_new}{numeric, magnetic field strength of new data}

\item{bw_ref}{numeric, bandwidth of reference}

\item{bw_new}{numeric, bandwidth of new data}

\item{dp_ref}{numeric, number of data points measured for reference}

\item{dp_new}{numeric, number of data points measured for new data}
}
\value{
new_data, matrix, nrow(data) x dp_ref
}
\description{
Adjust measurements for magnetic field strength, bandwidth and number of
measured data points with respect to a reference dataset.
}
