% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align_spectra}
\alias{align_spectra}
\title{align spectra}
\usage{
align_spectra(spectra, region, ref_pos = NULL)
}
\arguments{
\item{spectra}{matrix containing NMR spectra in rows}

\item{region}{numeric vector containing indices of a consecutive region of features, e.g. 1:100}

\item{ref_pos}{optional, numeric, the new index of the highest peak in the specified region.
If not supplied, the minimum index of the peaks in all samples in the set is used.}
}
\value{
matrix containing the shifted spectra
}
\description{
align all spectra in a data set to the highest peak in a region.
In this package, that is used to align all spectra to the choline peak.
}
