% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{bin_spectra}
\alias{bin_spectra}
\title{binning of features}
\usage{
bin_spectra(spectra, bin_size)
}
\arguments{
\item{spectra}{matrix containing NMR spectra in rows}

\item{bin_size}{numeric, number of features to combine in one bin (by averaging)}
}
\value{
binned_data
}
\description{
Take a matrix and reduce the number of features by using bins
}
