% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runWebApp}
\alias{runWebApp}
\title{Run Shiny Application for prediction}
\usage{
runWebApp(...)
}
\arguments{
\item{...}{parameters passed to shiny::shinyApp()}
}
\description{
This starts the shiny app.
By default, the server will be started on the default shiny port.
If a default browser is set on the system, the
page should open automatically.
}
\details{
Advanced users may pass arguments that will be passed to shiny::shinyApp() for
configuration of the server adress and port
}
