% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runWebApp}
\alias{runWebApp}
\title{Run Shiny Application for prediction}
\usage{
runWebApp(port = 8080, launch.browser = TRUE)
}
\arguments{
\item{port}{the port on which the server is run, default 8080}

\item{launch.browser}{bool, should default browser be launched to display the app? default TRUE}
}
\description{
This starts the shiny app.
}
